Change frontend API to query ES
Documents are now retrieved from the Elasticsearch database
instead of a file. Search results still come from a file,
but their individual records are retrieved from the database.

Assumes that the ES server is running on localhost:9200,
that an index called "article_test" exists and contains
documents of the type "article".

These assumptions will have to be changed later, but they're
sufficient for the demonstration tomorrow.